+++ CHANGELOG.md
@@ -17,3 +17,4 @@ Second version after feedback
 - Update: Naprawy view, add ordering, fresh view, configuration for statuses, saving state between views, more columns to display, more actions, and context window with actions.
 - Update: list component add new row on Enter click & disable Enter submit in request edition
 - Feature: notes for request & client notes
+- Feature: producer names switch and configuration